Windows fix for extracellular CRxD in python2.7.

+++ src/nrnpython/grids.c
@@ -214,8 +214,8 @@ void set_grid_concentrations(int grid_list_index, int index_in_list, PyObject* g
     */
     /* TODO: note that these will need updating anytime the structure of the model changes... look at the structure change count at each advance and trigger a callback to regenerate if necessary */
     Grid_node* g;
-    Py_ssize_t i;
-    Py_ssize_t n = PyList_Size(grid_indices);
+    int i;
+    int n = (int)PyList_Size(grid_indices);
 
     /* Find the Grid Object */
     g = Parallel_grids[grid_list_index];
@@ -249,8 +249,8 @@ void set_grid_currents(int grid_list_index, int index_in_list, PyObject* grid_in
     */
     /* TODO: note that these will need updating anytime the structure of the model changes... look at the structure change count at each advance and trigger a callback to regenerate if necessary */
     Grid_node* g;
-    Py_ssize_t i;
-    Py_ssize_t n = PyList_Size(grid_indices);
+    int i;
+    int n = (int)PyList_Size(grid_indices);
     long* dests;
 
     /* Find the Grid Object */
